refactor: group components by modules
