+++ src/modules/converter/CodecSelect/CodecSelect.js
@@ -1,13 +1,13 @@
 import React, { useCallback } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import Select from '../Select';
-import codecSetting from '../../config/conversion/audio/codec';
-import { getSingleConversionSetting } from '../../store/conversionSettings/conversionSettings.selectors';
+import Select from '../../../components/Select';
+import codecSetting from '../../../config/conversion/audio/codec';
+import { getSingleConversionSetting } from '../../../store/conversionSettings/conversionSettings.selectors';
 import {
     resetConversionSettings,
     setConversionSetting,
-} from '../../store/conversionSettings/conversionSettings.actions';
-import { CONVERSION_SETTINGS } from '../../store/conversionSettings/conversionSettings.constants';
+} from '../../../store/conversionSettings/conversionSettings.actions';
+import { CONVERSION_SETTINGS } from '../../../store/conversionSettings/conversionSettings.constants';
 
 const CodeSelect = () => {
     const dispatch = useDispatch();
