+++ src/modules/video-comparison/Player/index.js
